+++ core/src/game_state/mod.rs
@@ -1,8 +1,7 @@
 use std::collections::HashSet;
-use std::ops::{Deref, DerefMut};
+use std::ops::Deref;
 
 use anyhow::{anyhow, bail, Error};
-use rand::{seq::SliceRandom, RngCore};
 use serde::{Deserialize, Serialize};
 
 use crate::bidding::Bid;
@@ -11,12 +10,14 @@ use crate::hands::Hands;
 use crate::message::MessageVariant;
 use crate::settings::{
     FirstLandlordSelectionPolicy, Friend, FriendSelection, FriendSelectionPolicy, GameMode,
-    GameModeSettings, GameStartPolicy, KittyBidPolicy, KittyTheftPolicy, PropagatedState,
+    KittyBidPolicy, KittyTheftPolicy, PropagatedState,
 };
-use crate::types::{Card, Number, PlayerID, Rank, Trump, ALL_SUITS};
+use crate::types::{Card, Number, PlayerID, Rank, Trump};
 
+pub mod initialize_phase;
 pub mod play_phase;
 
+use initialize_phase::InitializePhase;
 use play_phase::PlayPhase;
 
 macro_rules! bail_unwrap {
@@ -106,9 +107,6 @@ impl GameState {
         }
     }
 
-    // Note: Clippy bug here which incorrectly triggers on .map(|()| vec![]):
-    // https://github.com/rust-lang/rust-clippy/issues/7224
-    #[allow(clippy::redundant_closure)]
     pub fn kick(&mut self, id: PlayerID) -> Result<Vec<MessageVariant>, Error> {
         match self {
             GameState::Initialize(ref mut p) => p.remove_player(id),
@@ -542,7 +540,7 @@ impl ExchangePhase {
         let mut propagated = self.propagated.clone();
         msgs.extend(propagated.make_all_observers_into_players()?);
 
-        Ok((InitializePhase { propagated }, msgs))
+        Ok((InitializePhase::from_propagated(propagated), msgs))
     }
 }
 
@@ -803,234 +801,7 @@ impl DrawPhase {
         let mut propagated = self.propagated.clone();
         msgs.extend(propagated.make_all_observers_into_players()?);
 
-        Ok((InitializePhase { propagated }, msgs))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct InitializePhase {
-    propagated: PropagatedState,
-}
-
-impl InitializePhase {
-    pub fn new() -> Self {
-        Self {
-            propagated: PropagatedState::default(),
-        }
-    }
-
-    pub fn propagated(&self) -> &PropagatedState {
-        &self.propagated
-    }
-
-    pub fn propagated_mut(&mut self) -> &mut PropagatedState {
-        &mut self.propagated
-    }
-
-    pub fn start(&self, id: PlayerID) -> Result<DrawPhase, Error> {
-        if self.propagated.players.len() < 4 {
-            bail!("not enough players")
-        }
-
-        if self.propagated.game_start_policy == GameStartPolicy::AllowLandlordOnly
-            && self.propagated.landlord.map(|l| l != id).unwrap_or(false)
-        {
-            bail!("Only the landlord can start the game")
-        }
-
-        let game_mode = match self.propagated.game_mode {
-            GameModeSettings::FindingFriends {
-                num_friends: Some(num_friends),
-                ..
-            } if num_friends + 1 < self.propagated.players.len() => GameMode::FindingFriends {
-                num_friends,
-                friends: vec![],
-            },
-            GameModeSettings::FindingFriends { .. } => GameMode::FindingFriends {
-                num_friends: (self.propagated.players.len() / 2) - 1,
-                friends: vec![],
-            },
-            GameModeSettings::Tractor if self.propagated.players.len() % 2 == 0 => {
-                GameMode::Tractor
-            }
-            GameModeSettings::Tractor => {
-                bail!("can only play tractor with an even number of players")
-            }
-        };
-
-        let mut rng = rand::thread_rng();
-
-        let position = self
-            .propagated
-            .landlord
-            .and_then(|landlord| {
-                self.propagated
-                    .players
-                    .iter()
-                    .position(|p| p.id == landlord)
-            })
-            .unwrap_or(rng.next_u32() as usize % self.propagated.players.len());
-
-        let level = if self.propagated.landlord.is_some() {
-            Some(self.propagated.players[position].rank())
-        } else {
-            None
-        };
-
-        let num_decks = self.propagated.num_decks();
-        if num_decks == 0 {
-            bail!("need at least one deck to start the game");
-        }
-        let decks = self.propagated.decks()?;
-        let mut deck = Vec::with_capacity(decks.iter().map(|d| d.len()).sum::<usize>());
-        for deck_ in &decks {
-            deck.extend(deck_.cards());
-        }
-        // Ensure that it is possible to bid for the landlord, if set, or all players, if not.
-        match level {
-            Some(Rank::Number(level)) if decks.iter().any(|d| d.includes_number(level)) => (),
-            Some(Rank::NoTrump) => (),
-            None if self.players.iter().all(|p| {
-                decks.iter().any(|d| match p.level {
-                    Rank::Number(level) => d.includes_number(level),
-                    Rank::NoTrump => true,
-                })
-            }) => {}
-            _ => bail!("deck configuration is missing cards needed to bid"),
-        }
-
-        deck.shuffle(&mut rng);
-
-        let mut removed_cards = vec![];
-
-        let kitty_size = match self.propagated.kitty_size {
-            Some(size)
-                if deck.len() % self.propagated.players.len()
-                    == size % self.propagated.players.len() =>
-            {
-                size
-            }
-            Some(size) => {
-                // Remove cards from the deck, until the deck and kitty together work out to the
-                // appropriate number of cards.
-                let num_players = self.propagated.players.len();
-
-                let min_number: Number = decks
-                    .iter()
-                    .map(|d| d.min)
-                    .min()
-                    .ok_or_else(|| anyhow!("no minimum value in deck?"))?;
-
-                // Choose a card to remove that doesn't unfairly disadvantage a particular player,
-                // and ideally isn't points either.
-                let removed_card_number = match level {
-                    Some(Rank::Number(level)) if level == min_number => {
-                        // If the minimum value isn't an A, this will be reasonable, otherwise
-                        // it'll remove a trump card from the deck...
-                        min_number.successor().unwrap_or(min_number)
-                    }
-                    Some(_) => min_number,
-                    None => {
-                        let mut bad_levels = self
-                            .propagated
-                            .players
-                            .iter()
-                            .flat_map(|p| match p.level {
-                                Rank::Number(n) => Some(n),
-                                Rank::NoTrump => None,
-                            })
-                            .collect::<HashSet<Number>>();
-                        bad_levels.insert(Number::Five);
-                        bad_levels.insert(Number::Ten);
-                        bad_levels.insert(Number::King);
-                        let mut n = min_number;
-                        loop {
-                            if !bad_levels.contains(&n) {
-                                break n;
-                            }
-                            n = match n.successor() {
-                                Some(nn) => nn,
-                                // If we somehow have enough players that we can't remove cards
-                                // without disadvantaging _someone_, or choosing points,
-                                // arbitrarily choose to remove twos.
-                                None => break min_number,
-                            };
-                        }
-                    }
-                };
-
-                let mut suit_idx = ALL_SUITS.len() - 1;
-
-                while deck.len() % num_players != size % num_players {
-                    let card_to_remove = Card::Suited {
-                        suit: ALL_SUITS[suit_idx],
-                        number: removed_card_number,
-                    };
-                    suit_idx = if suit_idx == 0 {
-                        ALL_SUITS.len() - 1
-                    } else {
-                        suit_idx - 1
-                    };
-
-                    // Attempt to remove the card from the deck.
-                    match deck.iter().position(|c| *c == card_to_remove) {
-                        Some(idx) => {
-                            deck.remove(idx);
-                            removed_cards.push(card_to_remove);
-                        }
-                        // Note: we would only hit this case if there are fewer decks than players,
-                        // which should be prevented in the settings layer.
-                        None => bail!(format!(
-                            "Couldn't find {:?} in the deck to remove",
-                            card_to_remove
-                        )),
-                    }
-                }
-                size
-            }
-            None => {
-                let mut kitty_size = deck.len() % self.propagated.players.len();
-                if kitty_size == 0 {
-                    kitty_size = self.propagated.players.len();
-                }
-                if kitty_size < 5 {
-                    kitty_size += self.propagated.players.len();
-                }
-                kitty_size
-            }
-        };
-
-        let propagated = self.propagated.clone();
-
-        Ok(DrawPhase {
-            deck: (&deck[0..deck.len() - kitty_size]).to_vec(),
-            kitty: (&deck[deck.len() - kitty_size..]).to_vec(),
-            hands: Hands::new(self.propagated.players.iter().map(|p| p.id)),
-            bids: Vec::new(),
-            revealed_cards: 0,
-            autobid: None,
-            propagated,
-            position,
-            num_decks,
-            decks,
-            game_mode,
-            level,
-            removed_cards,
-        })
-    }
-}
-
-impl Deref for InitializePhase {
-    type Target = PropagatedState;
-
-    fn deref(&self) -> &PropagatedState {
-        &self.propagated
-    }
-}
-
-impl DerefMut for InitializePhase {
-    fn deref_mut(&mut self) -> &mut PropagatedState {
-        &mut self.propagated
+        Ok((InitializePhase::from_propagated(propagated), msgs))
     }
 }
 
@@ -1041,7 +812,7 @@ mod tests {
         KittyTheftPolicy,
     };
 
-    use crate::game_state::{play_phase::PlayPhase, InitializePhase};
+    use crate::game_state::{initialize_phase::InitializePhase, play_phase::PlayPhase};
     use crate::message::MessageVariant;
     use crate::player::Player;
     use crate::types::{cards, Card, Number, PlayerID, Rank, FULL_DECK};
@@ -2056,7 +1827,7 @@ mod tests {
         play.finish_trick().unwrap();
 
         if let Ok((phase, _, _msgs)) = play.finish_game() {
-            assert_eq!(phase.propagated.landlord, Some(p3));
+            assert_eq!(phase.propagated().landlord, Some(p3));
         };
     }
 
@@ -2227,9 +1998,9 @@ mod tests {
 
         assert_eq!(
             new_init_phase
-                .propagated
-                .players
-                .into_iter()
+                .propagated()
+                .players()
+                .iter()
                 .map(|p| p.level)
                 .collect::<Vec<Rank>>(),
             vec![
